requirements.txt
# EEG Dashboard Dependencies

# Core scientific computing
numpy>=1.24.0
scipy>=1.10.0

# EEG-specific processing (industry standard)
mne>=1.6.0

# Real-time plotting & UI
PyQt6>=6.5.0
pyqtgraph>=0.13.0

# Serial communication with USB adapter
pyserial>=3.5

# Data export formats
pyedflib>=0.1.30  # For EDF/EDF+ format (standard EEG format)

# Configuration management
pyyaml>=6.0

# Optional: for advanced signal processing
# scikit-learn>=1.3.0  # For artifact detection ML models
